requirements.txt
azure-storage-blob==12.13
python-dotenv